backend/requirements.txt
opencv-python==4.10.0.84
numpy>=1.23.5,<2.0.0
tensorflow==2.17.0
face-recognition==1.3.0
boto3==1.35.44
openai==1.52.0
pillow==11.0.0
huggingface-hub==0.26.1
tqdm==4.66.5
mediapipe==0.10.15
python-ffmpeg==2.0.12